--- example/lib/demo3.dart
@@ -1,49 +1,0 @@
-/*
- * This file is part of Flutter-Sound (Flauto).
- *
- *   Flutter-Sound (Flauto) is free software: you can redistribute it and/or modify
- *   it under the terms of the Lesser GNU General Public License
- *   version 3 (LGPL3) as published by the Free Software Foundation.
- *
- *   Flutter-Sound (Flauto) is distributed in the hope that it will be useful,
- *   but WITHOUT ANY WARRANTY; without even the implied warranty of
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *   GNU General Public License for more details.
- *
- *   You should have received a copy of the Lesser GNU General Public License
- *   along with Flutter-Sound (Flauto).  If not, see <https://www.gnu.org/licenses/>.
- */
-
-import 'package:flutter/material.dart';
-
-import 'demo_util/demo3_body.dart';
-
-/// demonstrates the recording widget linked to a playback widget.
-void main() {
-  runApp(MyApp());
-}
-
-/// Example app.
-class MyApp extends StatefulWidget {
-  @override
-  _MyAppState createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Flutter Sound'),
-        ),
-        body: MainBody(),
-      ),
-    );
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-  }
-}
